+++ controllers/categoryController.js
@@ -1,7 +1,7 @@
-const { newProductsService } = require("../services");
+const { categoryService } = require("../services");
 
 const getNewProductsList = async (req, res) => {
-    const getNewProducts = await newProductsService.getNewProductsList
+    const getNewProducts = await categoryService.getNewProductsList()
     res.status(201).json({ getNewProducts })
 }
