+++ models/index.js
@@ -3,7 +3,7 @@ const reviewDao = require("./reviewDao");
 const productDao = require("./productDao");
 const userDao = require('./userDao')
 const cartDao = require("./cartDao")
-const newProductsDao = require("./newProductsDao")
+const categoryDao = require("./categoryDao")
 
 
 module.exports = {
@@ -12,5 +12,5 @@ module.exports = {
     productDao,
     userDao,
     cartDao,
-    newProductsDao
+    categoryDao
 }
