+++ routes/categoryRouter.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const { categoryController } = require('../controllers');
+
+const router = express.Router();
+
+router.get('/new', categoryController.getNewProductsList);
+
+module.exports = router;
