+++ routes/index.js
@@ -4,13 +4,13 @@ const router = express.Router();
 const reviewRouter = require("./reviewRouter");
 const userRouter = require('./userRouter');
 const cartRouter = require("./cartRouter")
-const newProductsRouter = require("./newProductsRouter");
+const categoryRouter = require("./categoryRouter");
 
 
 router.use('/users', userRouter);
 router.use("/review", reviewRouter);
 router.use("/cart", cartRouter);
-router.use("/category", newProductsRouter);
+router.use("/category", categoryRouter);
 
 
 module.exports = router
