--- routes/newProductsRouter.js
@@ -1,8 +1,0 @@
-const express = require('express');
-const { newProductsController } = require('../controllers');
-
-const router = express.Router();
-
-router.get('/new', newProductsController.newProducts);
-
-module.exports = router;
