+++ services/categoryService.js
@@ -1,0 +1,9 @@
+const { categoryDao } = require("../models");
+
+const getNewProductsList = async() => {
+    return await categoryDao.getNewProductsList()
+}
+
+module.exports = {
+    getNewProductsList
+}
