--- services/newProductsService.js
@@ -1,10 +1,0 @@
-const { newProductsDao } = require("../models");
-
-const getNewProductsList = async() => {
-    const newProducts = await newProductsDao.getNewProductsList()
-    return newProducts
-}
-
-module.exports = {
-    getNewProductsList
-}
